requirements.txt
# common
numpy==1.22.4
pandas==1.3.5
scikit-learn==1.5.2
torch==2.2.1+cu121
tqdm==4.67.1

# irg
accelerate==0.34.2
category-encoders==2.0.0
cupy-cuda12x==13.4.1
dython==0.7.5
joblib==1.4.2
ortools==9.12.4544
pot==0.9.5
psutil==5.9.8
realtabformer==0.1.5
scipy==1.13.1
transformers==4.39.0
xgboost==2.1.0

# baseline (other than baseline requirements.txt)
json-tricks==3.17.3
sdv==1.16.2

# evaluation
lightgbm==4.6.0
matplotlib==3.8.3
scipy==1.13.1
seaborn==0.13.2
PyYAML==6.0.2
xgboost==2.1.0
